gcc 4.8.x and -Wconversion.
svn:sync-xref-src-repo-rev: r110077

+++ src/VBox/Main/include/Global.h
@@ -1,4 +1,4 @@
-/* $Id: Global.h 109041 2016-07-22 18:36:43Z knut.osmundsen@oracle.com $ */
+/* $Id: Global.h 110077 2016-08-14 00:40:45Z knut.osmundsen@oracle.com $ */
 /** @file
  * VirtualBox COM API - Global Declarations and Definitions.
  */
@@ -82,7 +82,7 @@ public:
     };
 
     static const OSType sOSTypes[];
-    static uint32_t cOSTypes;
+    static size_t cOSTypes;
 
     /**
      * Maps VBOXOSTYPE to the OS type which is used in VM configs.
